requirements.txt
Django==4.0.4
django-cors-headers==3.11.0
django-filter==21.1
django-redis==5.2.0
djangorestframework==3.13.1
djangorestframework-simplejwt==5.1.0
gunicorn==19.9.0
psycopg2-binary==2.9.3
PyJWT==2.3.0
pyparsing==3.0.8
requests==2.27.1
sqlparse==0.4.2
urllib3==1.26.9
whitenoise==6.0.0
